Set single file on db module
