--- packages/brain/src/modules/db/validate.ts
@@ -1,2 +1,0 @@
-// To be implemented
-export const b = "";
